Moved logic over to water-trading-scenario since it's specific to that component. Added authentication guards to the paths
