+++ Source/Rio.Web/src/app/app-routing.module.ts
@@ -71,9 +71,9 @@ const routes: Routes = [
   { path: "manager-dashboard", component: ManagerDashboardComponent, canActivate: [UnauthenticatedAccessGuard, ManagerOnlyGuard, AcknowledgedDisclaimerGuard] },
   { path: "invite-user/:userID", component: UserInviteComponent, canActivate: [UnauthenticatedAccessGuard, ManagerOnlyGuard, AcknowledgedDisclaimerGuard] },
   { path: "invite-user", component: UserInviteComponent, canActivate: [UnauthenticatedAccessGuard, ManagerOnlyGuard, AcknowledgedDisclaimerGuard] },
-  { path: "about-groundwater-evaluation", component: AboutGroundwaterEvaluationComponent},
-  { path: "managed-recharge-scenario", component: ManagedRechargeScenarioComponent},
-  { path: "water-trading-scenario", component: WaterTradingScenarioComponent},
+  { path: "about-groundwater-evaluation", component: AboutGroundwaterEvaluationComponent, canActivate: [UnauthenticatedAccessGuard]},
+  { path: "managed-recharge-scenario", component: ManagedRechargeScenarioComponent, canActivate: [UnauthenticatedAccessGuard]},
+  { path: "water-trading-scenario", component: WaterTradingScenarioComponent, canActivate: [UnauthenticatedAccessGuard]},
   { path: "", component: HomeIndexComponent},
   { path: "disclaimer", component: DisclaimerComponent },
   { path: "disclaimer/:forced", component: DisclaimerComponent },
