+++ Source/Rio.Web/src/app/pages/water-trading-scenario/water-trading-scenario.component.html
@@ -22,7 +22,6 @@
         </div>
     </div>
     <div class="card-body table-responsive">
-        <scenario-map mapID="waterTradingScenarioMap" mapHeight="600px">
-        </scenario-map>
+        <div id={{mapID}} class="location-card" [style.height]="mapHeight"></div>
     </div>
 </div>
